Revert "ByFTP 1.0.12 — final premium UX, English-first and modular cleanup"
