--- internal/i18n/action_catalogs.go
@@ -1,72 +1,0 @@
-package i18n
-
-// actionCatalogs keeps dense workflow/dialog text out of the core catalog
-// files. Every key is supplied for every supported locale so switching
-// language never falls back to a mixed-language Windows action flow.
-var actionCatalogs = map[string]map[string]string{
-	"en": {
-		"action.failed_retry": "The action did not finish. Try again.",
-		"action.failed_title": "Action failed",
-		"file.choose_folder_failed_title": "Could not select folder", "file.choose_folder_failed_body": "The folder cannot be selected right now.",
-		"local.open_failed": "The local folder cannot be opened.", "remote.open_failed": "The server folder cannot be opened.",
-		"item.open_unsafe": "This item cannot be opened safely.", "item.symlink_download_blocked": "The symbolic link was not downloaded automatically for safety.", "item.save_unsafe": "This item cannot be saved safely on the local computer.",
-		"file.new_name": "New name:", "file.folder_name": "Folder name:", "file.default_folder": "New folder", "selection.items": "%d selected items",
-		"local.mkdir.title": "ByFTP — New local folder", "local.mkdir.failed_title": "Folder was not created", "local.mkdir.failed_body": "The folder cannot be created.", "local.mkdir.success": "Local folder created: %s",
-		"local.rename.select_one": "Select exactly one local item to rename.", "local.rename.title": "ByFTP — Rename local item", "local.rename.failed_title": "Rename failed", "local.rename.failed_body": "The item cannot be renamed.", "local.rename.success": "Local item renamed.",
-		"local.delete.select": "Select one or more local items to delete.", "local.delete.title": "ByFTP — Delete", "local.delete.question": "Delete selected local items?", "local.delete.warning": "This action cannot be undone through ByFTP.", "local.delete.partial_title": "Not all items were deleted", "local.delete.partial_body": "Some selected items could not be deleted.", "local.delete.status": "Local items deleted: %d • failed: %d",
-		"remote.mkdir.title": "ByFTP — New server folder", "remote.mkdir.progress": "Creating folder", "remote.mkdir.success": "Folder created: %s",
-		"remote.rename.select_one": "Select exactly one server item to rename.", "remote.rename.title": "ByFTP — Rename on server", "remote.rename.progress": "Renaming", "remote.rename.success": "Server item renamed.",
-		"remote.delete.select": "Select one or more server items to delete.", "remote.delete.title": "ByFTP — Delete on server", "remote.delete.question": "Delete selected items?", "remote.delete.warning": "Deletion on the server can be irreversible.", "remote.delete.progress": "Deleting", "remote.delete.success_prefix": "Server items deleted: ",
-		"permissions.select": "Select one or more items to change permissions.", "permissions.limit": "Select no more than 1000 items for one permission change.", "permissions.links_only": "Permissions were not changed: symbolic links are skipped for safety.", "permissions.title": "ByFTP — Permissions", "permissions.prompt": "Permissions for selected items, for example 644 or 755:", "permissions.progress": "Changing permissions", "permissions.success_prefix": "Permissions changed for items: ",
-		"remote.action.title": "ByFTP — Server", "remote.action.failed": "%s failed", "remote.action.body": "The server action cannot be completed.", "remote.batch.partial": "%s did not finish for all items", "remote.batch.body": "Some selected items could not be changed.", "batch.failed": " • failed: %d", "batch.skipped_links": " • links skipped: %d",
-		"transfer.links_skipped": "Selected symbolic links were not transferred for safety.", "transfer.none": "There are no items to transfer.", "transfer.adding_selection": "Adding selected items to the transfer queue…", "transfer.partial_title": "Not all selected items were queued", "transfer.partial_body": "Some selected items could not be added to the transfer queue.", "transfer.queue_files": "Queued: %d files", "transfer.queue_dirs": ", %d folders", "transfer.queue_failed": " • selections failed: %d", "transfer.queue_skipped_links": " • links skipped: %d",
-		"transfer.select_upload": "Select one or more local files or folders to upload.", "transfer.select_download": "Select one or more server files or folders to download.", "transfer.tree_preparing": "Preparing full-folder transfer…", "transfer.tree_title": "ByFTP — Folder transfer", "transfer.tree_failed_title": "Folder was not queued", "transfer.tree_failed_body": "The folder transfer cannot be started.", "transfer.tree_added": "Folder queued: %d files, %d folders.",
-		"connection.lost_stopping": "The connection is no longer available. Stopping transfers…", "connection.lost_done": "The connection is no longer available. Active transfers were stopped; the server and username remain entered.",
-		"transfer.adding": "Adding transfer…", "transfer.start_failed_title": "Transfer was not started", "transfer.start_failed_body": "Check the connection and selected files.", "transfer.start_failed_status": "The transfer cannot be started.", "transfer.queued": "Transfer added to the queue.",
-		"transfer.pause_status": "Transfer queue paused. Active transfers may finish their current work; new work waits until resume.", "transfer.resume_status": "Transfer queue resumed.", "transfer.clear_status": "Finished transfers removed from the queue.",
-		"transfer.select_active_cancel": "Select one or more active transfers to cancel.", "transfer.cancel_failed_body": "The selected transfers cannot be cancelled right now.", "transfer.cancelled_count": "Transfers cancelled: %d", "transfer.select_retry": "Select one or more failed or cancelled transfers to retry.", "transfer.retry_requires_connection": "Connect to the server before retrying transfers.", "transfer.retry_failed_body": "The selected transfers cannot be retried right now.", "transfer.requeued_count": "Requeued: %d",
-		"profile.edit_blocked_title": "Profile cannot be changed while connected", "profile.edit_blocked_body": "Wait for the connection attempt to finish or disconnect before saving profile changes.", "profile.dialog_title": "ByFTP — Profile", "profile.name_prompt": "Profile name:", "profile.name_missing_title": "Profile name is required", "profile.name_missing_body": "Enter a name that identifies this connection.",
-		"privacy.title": "ByFTP — Privacy", "profile.store_credentials_title": "Save credentials on this computer?", "profile.store_credentials_body": "Yes = entered credentials are stored in the ByFTP vault protected by Windows.\nNo = the profile is saved without a stored password or private-key passphrase.", "profile.retain_credentials_title": "Keep saved credentials?", "profile.retain_credentials_body": "This profile already contains saved credentials that still belong to this identity.\n\nYes = keep them.\nNo = remove them from the ByFTP vault.", "profile.credentials_auto_removed_intro": "Credentials that no longer belong to the new server, user, or private key will be removed automatically.\n\n", "profile.old_credentials_title": "Old credentials will not be carried over", "profile.old_credentials_body": "The server, port, username, or private key changed. For safety, old saved credentials are removed from this profile. Enter them again if you want to save credentials for the new identity.",
-		"profile.save_failed_title": "Profile was not saved", "profile.save_failed_body": "The profile cannot be saved right now.", "profile.saved": "Profile saved: %s", "profile.none_selected_title": "No saved profile selected", "profile.none_selected_body": "Select the profile you want to delete.", "profile.delete_connected_title": "Profile cannot be deleted while connected", "profile.delete_connected_body": "Disconnect first.", "profile.delete_title": "ByFTP — Profiles", "profile.delete_question": "Delete profile?", "profile.delete_body": "%s\n\nSaved credentials will be removed from the ByFTP vault.", "profile.delete_failed_title": "Profile was not deleted", "profile.delete_failed_body": "The profile cannot be deleted right now.", "profile.deleted": "Profile deleted.",
-	},
-	"hr": {
-		"action.failed_retry": "Radnja nije dovršena. Pokušajte ponovno.", "action.failed_title": "Radnja nije uspjela", "file.choose_folder_failed_title": "Odabir mape nije uspio", "file.choose_folder_failed_body": "Mapu trenutačno nije moguće odabrati.", "local.open_failed": "Lokalnu mapu nije moguće otvoriti.", "remote.open_failed": "Mapu na poslužitelju nije moguće otvoriti.", "item.open_unsafe": "Ovu stavku nije moguće sigurno otvoriti.", "item.symlink_download_blocked": "Simbolička poveznica nije automatski preuzeta radi sigurnosti.", "item.save_unsafe": "Ovu stavku nije moguće sigurno spremiti na lokalno računalo.", "file.new_name": "Novi naziv:", "file.folder_name": "Naziv nove mape:", "file.default_folder": "Nova mapa", "selection.items": "%d odabranih stavki", "local.mkdir.title": "ByFTP — Nova lokalna mapa", "local.mkdir.failed_title": "Mapa nije stvorena", "local.mkdir.failed_body": "Mapu nije moguće stvoriti.", "local.mkdir.success": "Lokalna mapa stvorena: %s", "local.rename.select_one": "Za preimenovanje odaberite točno jednu lokalnu stavku.", "local.rename.title": "ByFTP — Preimenuj lokalno", "local.rename.failed_title": "Preimenovanje nije uspjelo", "local.rename.failed_body": "Stavku nije moguće preimenovati.", "local.rename.success": "Lokalna stavka preimenovana.", "local.delete.select": "Odaberite jednu ili više lokalnih stavki za brisanje.", "local.delete.title": "ByFTP — Brisanje", "local.delete.question": "Obrisati odabrane lokalne stavke?", "local.delete.warning": "Ova radnja se ne može poništiti kroz ByFTP.", "local.delete.partial_title": "Nisu obrisane sve stavke", "local.delete.partial_body": "Dio odabranih stavki nije moguće obrisati.", "local.delete.status": "Obrisano lokalnih stavki: %d • neuspjelo: %d", "remote.mkdir.title": "ByFTP — Nova mapa na poslužitelju", "remote.mkdir.progress": "Stvaranje mape", "remote.mkdir.success": "Mapa stvorena: %s", "remote.rename.select_one": "Za preimenovanje odaberite točno jednu udaljenu stavku.", "remote.rename.title": "ByFTP — Preimenuj na poslužitelju", "remote.rename.progress": "Preimenovanje", "remote.rename.success": "Udaljena stavka preimenovana.", "remote.delete.select": "Odaberite jednu ili više udaljenih stavki za brisanje.", "remote.delete.title": "ByFTP — Brisanje na poslužitelju", "remote.delete.question": "Obrisati odabrane stavke?", "remote.delete.warning": "Brisanje na poslužitelju može biti nepovratno.", "remote.delete.progress": "Brisanje", "remote.delete.success_prefix": "Obrisano udaljenih stavki: ", "permissions.select": "Odaberite jednu ili više stavki za promjenu dozvola.", "permissions.limit": "Za jednu promjenu dozvola odaberite najviše 1000 stavki.", "permissions.links_only": "Dozvole nisu mijenjane: simboličke poveznice preskaču se radi sigurnosti.", "permissions.title": "ByFTP — Dozvole", "permissions.prompt": "Dozvole za odabrane stavke, npr. 644 ili 755:", "permissions.progress": "Promjena dozvola", "permissions.success_prefix": "Dozvole promijenjene za stavki: ", "remote.action.title": "ByFTP — Poslužitelj", "remote.action.failed": "%s nije uspjelo", "remote.action.body": "Radnju na poslužitelju nije moguće dovršiti.", "remote.batch.partial": "%s nije dovršeno za sve stavke", "remote.batch.body": "Dio odabranih stavki nije moguće promijeniti.", "batch.failed": " • neuspjelo: %d", "batch.skipped_links": " • preskočeno poveznica: %d", "transfer.links_skipped": "Odabrane poveznice nisu prenesene radi sigurnosti.", "transfer.none": "Nema stavki za prijenos.", "transfer.adding_selection": "Dodavanje odabranih stavki u red prijenosa…", "transfer.partial_title": "Nisu dodane sve odabrane stavke", "transfer.partial_body": "Dio odabranih stavki nije moguće dodati u prijenos.", "transfer.queue_files": "Dodano u red: %d datoteka", "transfer.queue_dirs": ", %d mapa", "transfer.queue_failed": " • neuspjelo odabira: %d", "transfer.queue_skipped_links": " • preskočeno poveznica: %d", "transfer.select_upload": "Odaberite jednu ili više lokalnih datoteka ili mapa za prijenos.", "transfer.select_download": "Odaberite jednu ili više datoteka ili mapa na poslužitelju za preuzimanje.", "transfer.tree_preparing": "Priprema prijenosa cijele mape…", "transfer.tree_title": "ByFTP — Prijenos mape", "transfer.tree_failed_title": "Mapa nije dodana u prijenos", "transfer.tree_failed_body": "Prijenos mape nije moguće pokrenuti.", "transfer.tree_added": "Mapa dodana: %d datoteka, %d mapa.", "connection.lost_stopping": "Veza više nije dostupna. Zaustavljanje prijenosa…", "connection.lost_done": "Veza više nije dostupna. Aktivni prijenosi su zaustavljeni; poslužitelj i korisničko ime ostali su uneseni.", "transfer.adding": "Dodavanje prijenosa…", "transfer.start_failed_title": "Prijenos nije pokrenut", "transfer.start_failed_body": "Provjerite vezu i odabrane datoteke.", "transfer.start_failed_status": "Prijenos nije moguće pokrenuti.", "transfer.queued": "Prijenos dodan u red čekanja.", "transfer.pause_status": "Red prijenosa pauziran. Aktivni prijenosi mogu dovršiti trenutačni rad; novi čekaju nastavak.", "transfer.resume_status": "Red prijenosa nastavljen.", "transfer.clear_status": "Završeni prijenosi uklonjeni iz reda.", "transfer.select_active_cancel": "Odaberite jedan ili više aktivnih prijenosa za otkazivanje.", "transfer.cancel_failed_body": "Odabrane prijenose trenutačno nije moguće otkazati.", "transfer.cancelled_count": "Otkazano prijenosa: %d", "transfer.select_retry": "Odaberite jedan ili više neuspjelih ili otkazanih prijenosa za ponavljanje.", "transfer.retry_requires_connection": "Povežite se s poslužiteljem prije ponavljanja prijenosa.", "transfer.retry_failed_body": "Odabrane prijenose trenutačno nije moguće ponoviti.", "transfer.requeued_count": "Ponovno dodano u red: %d", "profile.edit_blocked_title": "Profil nije moguće mijenjati tijekom veze", "profile.edit_blocked_body": "Pričekajte završetak povezivanja ili prekinite vezu pa spremite promjene profila.", "profile.dialog_title": "ByFTP — Profil", "profile.name_prompt": "Naziv profila:", "profile.name_missing_title": "Naziv profila nedostaje", "profile.name_missing_body": "Upišite naziv po kojem ćete prepoznati ovu vezu.", "privacy.title": "ByFTP — Privatnost", "profile.store_credentials_title": "Spremiti vjerodajnice na ovom računalu?", "profile.store_credentials_body": "Da = upisane vjerodajnice spremit će se u ByFTP spremište zaštićeno sustavom Windows.\nNe = profil će se spremiti bez spremljene lozinke i zaporke privatnog ključa.", "profile.retain_credentials_title": "Zadržati spremljene vjerodajnice?", "profile.retain_credentials_body": "Profil već sadrži spremljene vjerodajnice koje još pripadaju ovom identitetu.\n\nDa = zadrži ih.\nNe = ukloni ih iz ByFTP spremišta.", "profile.credentials_auto_removed_intro": "Vjerodajnice koje više ne pripadaju novom poslužitelju, korisniku ili privatnom ključu bit će automatski uklonjene.\n\n", "profile.old_credentials_title": "Stare vjerodajnice neće se prenijeti", "profile.old_credentials_body": "Promijenili ste poslužitelj, port, korisničko ime ili privatni ključ. Radi zaštite stare spremljene vjerodajnice uklanjaju se iz ovog profila. Ako ih želite spremiti za novi identitet, upišite ih ponovno.", "profile.save_failed_title": "Profil nije spremljen", "profile.save_failed_body": "Profil trenutačno nije moguće spremiti.", "profile.saved": "Profil spremljen: %s", "profile.none_selected_title": "Nije odabran spremljeni profil", "profile.none_selected_body": "Odaberite profil koji želite obrisati.", "profile.delete_connected_title": "Profil se ne briše tijekom veze", "profile.delete_connected_body": "Prvo prekinite vezu.", "profile.delete_title": "ByFTP — Profili", "profile.delete_question": "Obrisati profil?", "profile.delete_body": "%s\n\nSpremljene vjerodajnice bit će uklonjene iz ByFTP spremišta.", "profile.delete_failed_title": "Profil nije obrisan", "profile.delete_failed_body": "Profil trenutačno nije moguće obrisati.", "profile.deleted": "Profil obrisan.",
-	},
-}
-
-// Secondary action text is intentionally complete rather than falling back to
-// English. The compact translations below cover the same keys as English and
-// Croatian while keeping this module reviewable.
-func init() {
-	for code, extra := range actionCatalogs {
-		for key, value := range extra {
-			catalogs[code][key] = value
-		}
-	}
-	for _, code := range []string{"de", "fr", "es", "tr", "el", "pt", "zh", "ru", "hi", "ja"} {
-		if _, ok := actionCatalogs[code]; ok {
-			continue
-		}
-		actionCatalogs[code] = translateActionCatalog(code)
-		for key, value := range actionCatalogs[code] {
-			catalogs[code][key] = value
-		}
-	}
-}
-
-// translateActionCatalog provides curated per-locale workflow strings while
-// preserving format placeholders exactly. Keeping the translations grouped by
-// semantic key avoids duplicating the Windows control flow itself.
-func translateActionCatalog(code string) map[string]string {
-	base := actionCatalogs["en"]
-	out := make(map[string]string, len(base))
-	for key, value := range base {
-		out[key] = value
-	}
-	// For secondary locales, the core product controls/errors are already fully
-	// localized. Dense workflow prose remains readable English until reviewed by
-	// a native speaker rather than pretending an unverified machine translation
-	// is authoritative. Catalog completeness is preserved and no Croatian text
-	// leaks into non-Croatian locales.
-	return out
-}
